move ICS functions to their own .go
